requirements.txt
aiocsv==1.3.2
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
alembic==1.15.1
annotated-types==0.7.0
anyio==4.8.0
anywidget==0.9.13
asttokens==3.0.0
asyncpg==0.30.0
attrs==25.3.0
blinker==1.9.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
comm==0.2.2
contourpy==1.3.1
cycler==0.12.1
dash==2.18.2
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
decorator==5.1.1
environs==14.1.1
executing==2.2.0
fastapi==0.115.11
fastapi_cors==0.0.6
Flask==3.0.3
fonttools==4.56.0
frozenlist==1.5.0
gevent==24.11.1
greenlet==3.1.1
h11==0.14.0
idna==3.10
importlib_metadata==8.6.1
influxdb-client==1.48.0
ipython==8.32.0
ipywidgets==8.1.5
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.5
jupyterlab_widgets==3.0.13
kiwisolver==1.4.8
Mako==1.3.9
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib-inline==0.1.7
multidict==6.2.0
narwhals==1.25.2
nest-asyncio==1.6.0
numpy==2.2.2
packaging==24.2
parso==0.8.4
pillow==11.1.0
plotly==6.0.0
prompt_toolkit==3.0.50
propcache==0.3.1
psygnal==0.12.0
pure_eval==0.2.3
pycparser==2.22
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
Pygments==2.19.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
reactivex==4.0.4
redis==5.2.1
requests==2.32.3
retrying==1.3.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.39
stack-data==0.6.3
starlette==0.46.0
traitlets==5.14.3
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
wcwidth==0.2.13
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.0.6
widgetsnbextension==4.0.13
yarl==1.18.3
zipp==3.21.0
zope.event==5.0
zope.interface==7.2
